Separate the test of the scorer from the test of the ner_tagger itself
